requirements.txt
aws-cdk-lib==2.189.0
constructs>=10.0.0,<11.0.0
pyyaml
boto3
aws-cdk.aws-lambda-python-alpha
sagemaker>=2.198.0
xgboost